# backend/swagger/api-docs.yml
openapi: 3.0.0
info:
  title: ZeYang API
  description: ZeYang 房地產專案管理系統 API 文檔
  version: 2.1.0
  contact:
    name: API Support
    email: dev@ZeYang.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/api
    description: Development server
  - url: https://api.ZeYang.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: object
          properties:
            code:
              type: string
              example: ERROR_CODE
            details:
              type: string
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
...
    Contact:
      type: object
      properties:
        id:
          type: integer
          description: 聯絡 ID
        name:
          type: string
          description: 姓名
        email:
          type: string
          description: 電子郵件
        phone:
          type: string
          description: 電話
        company:
          type: string
          description: 公司
        subject:
          type: string
          description: 主旨
        message:
          type: string
          description: 訊息內容
        projectId:
          type: integer
          description: 關聯專案 ID
        source:
          type: string
          description: 來源
        isRead:
          type: boolean
          description: 是否已讀
        isReplied:
          type: boolean
          description: 是否已回覆
        created_at:
          type: string
          format: date-time
          description: 建立時間
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 總筆數
        page:
          type: integer
          description: 目前頁數
        limit:
          type: integer
          description: 每頁筆數
        pages:
          type: integer
          description: 總頁數
        hasNext:
          type: boolean
          description: 是否有下一頁
        hasPrev:
          type: boolean
          description: 是否有上一頁
paths:
...
  /contacts:
    get:
      tags:
        - Contacts
      summary: 取得聯絡訊息列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: isRead
          schema:
            type: boolean
        - in: query
          name: isReplied
          schema:
            type: boolean
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date
        - in: query
          name: dateTo
          schema:
            type: string
            format: date
      responses:
        200:
          description: 聯絡訊息列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Contact'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
    
    post:
      tags:
        - Contacts
      summary: 提交聯絡表單
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - message
              properties:
                name:
                  type: string
                  maxLength: 100
                email:
                  type: string
                  format: email
                  maxLength: 100
                phone:
                  type: string
                  maxLength: 50
                company:
                  type: string
                  maxLength: 100
                subject:
                  type: string
                  maxLength: 200
                message:
                  type: string
                  minLength: 10
                  maxLength: 2000
                source:
                  type: string
      responses:
        201:
          description: 表單提交成功
        422:
          $ref: '#/components/responses/ValidationError'
        429:
          description: 請求過於頻繁
  /contacts/statistics:
    get:
      tags:
        - Contacts
      summary: 取得聯絡統計資料
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
      responses:
        200:
          description: 統計資料
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  /contacts/{id}:
    get:
      tags:
        - Contacts
      summary: 取得單一聯絡訊息
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 聯絡訊息詳情
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Contacts
      summary: 刪除聯絡訊息
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 聯絡訊息刪除成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
  /contacts/{id}/read:
    put:
      tags:
        - Contacts
      summary: 標記為已讀
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 標記成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  /contacts/{id}/reply:
    put:
      tags:
        - Contacts
      summary: 回覆聯絡訊息
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: 回覆內容
                notes:
                  type: string
                  description: 內部備註
      responses:
        200:
          description: 回覆成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'

  # 設定管理端點
...
components:
  responses:
    Unauthorized:
      description: 未授權
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 權限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 資源未找到
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: 驗證錯誤
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
...